AMUSE_project_YESSIR/requirements.txt
amuse-framework
amuse-seba
amuse-sse
amuse-bhtree
amuse-fi
numpy
matplotlib
plotly